requirements.txt
Django==3.1.3
python-docx==0.8.10
requests
Pillow